theme providerの位置を変更
